refactor(contracts): extract shared messaging and RabbitMQ options (Cutube-vmv.2)
