+++ cutube.sln
@@ -23,6 +23,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Cutube.Worker", "src\Cutube
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Cutube.Worker.Tests", "tests\Cutube.Worker.Tests\Cutube.Worker.Tests.csproj", "{DAC6B9CC-8389-4835-B9D8-6644AC4ECA17}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Cutube.Contracts", "src\Cutube.Contracts\Cutube.Contracts.csproj", "{3967E688-B8B5-4CCD-B138-4133DAFF92B4}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -129,6 +131,18 @@ Global
 		{DAC6B9CC-8389-4835-B9D8-6644AC4ECA17}.Release|x64.Build.0 = Release|Any CPU
 		{DAC6B9CC-8389-4835-B9D8-6644AC4ECA17}.Release|x86.ActiveCfg = Release|Any CPU
 		{DAC6B9CC-8389-4835-B9D8-6644AC4ECA17}.Release|x86.Build.0 = Release|Any CPU
+		{3967E688-B8B5-4CCD-B138-4133DAFF92B4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{3967E688-B8B5-4CCD-B138-4133DAFF92B4}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{3967E688-B8B5-4CCD-B138-4133DAFF92B4}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{3967E688-B8B5-4CCD-B138-4133DAFF92B4}.Debug|x64.Build.0 = Debug|Any CPU
+		{3967E688-B8B5-4CCD-B138-4133DAFF92B4}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{3967E688-B8B5-4CCD-B138-4133DAFF92B4}.Debug|x86.Build.0 = Debug|Any CPU
+		{3967E688-B8B5-4CCD-B138-4133DAFF92B4}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{3967E688-B8B5-4CCD-B138-4133DAFF92B4}.Release|Any CPU.Build.0 = Release|Any CPU
+		{3967E688-B8B5-4CCD-B138-4133DAFF92B4}.Release|x64.ActiveCfg = Release|Any CPU
+		{3967E688-B8B5-4CCD-B138-4133DAFF92B4}.Release|x64.Build.0 = Release|Any CPU
+		{3967E688-B8B5-4CCD-B138-4133DAFF92B4}.Release|x86.ActiveCfg = Release|Any CPU
+		{3967E688-B8B5-4CCD-B138-4133DAFF92B4}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -141,5 +155,6 @@ Global
 		{9F9C26F0-68C6-4F6C-9D50-06364F5B87BA} = {0AB3BF05-4346-4AA6-1389-037BE0695223}
 		{73BAD635-6600-4752-8F9B-4C31A7444C58} = {827E0CD3-B72D-47B6-A68D-7590B98EB39B}
 		{DAC6B9CC-8389-4835-B9D8-6644AC4ECA17} = {0AB3BF05-4346-4AA6-1389-037BE0695223}
+		{3967E688-B8B5-4CCD-B138-4133DAFF92B4} = {827E0CD3-B72D-47B6-A68D-7590B98EB39B}
 	EndGlobalSection
 EndGlobal
