+++ src/Cutube.Worker/Services/DownloadProcessingService.cs
@@ -1,4 +1,4 @@
-using Cutube.Api.Queuing.Messages;
+using Cutube.Contracts.Messages;
 using Cutube.Worker.Configuration;
 using Cutube.Worker.Services.Models;
 using Microsoft.Extensions.Options;
